refactoring

+++ database/migrations/2022_08_22_194010_create_exchange_rates_models_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('exchange_rates_models', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('currency_id');
-            $table->unsignedBigInteger('buy');
-            $table->unsignedBigInteger('sale');
+            $table->double('buy');
+            $table->double('sale');
             $table->timestamps();
         });
     }
